lower accuracy to single precision floating point number to better utalize the ARM NEON SIMD of the RoboRIO (about 4x speedup on certain segments of the calculator)

+++ src/main/java/frc/robot/util/BallisticCalculator/BallisticCalculator.java
@@ -10,51 +10,51 @@ import java.util.stream.IntStream;
 public class BallisticCalculator {
 
     public enum BallisticCalculatorMode {
-        FAST(0.04,0.01),
-        ACCURATE(0.02,0.005);
+        FAST(0.04f,0.01f),
+        ACCURATE(0.02f,0.005f);
 
-        private double coarseDt, fineDt;
+        private float coarseDt, fineDt;
 
-        public double getCoarseDt() {
+        public float getCoarseDt() {
             return coarseDt;
         }
 
-        public double getFineDt() {
+        public float getFineDt() {
             return fineDt;
         }
 
-        BallisticCalculatorMode(double coarseDt, double fineDt) {
+        BallisticCalculatorMode(float coarseDt, float fineDt) {
             this.coarseDt = coarseDt;
             this.fineDt = fineDt;
         }
     }
 
-    private static double[] LinSpell(double start, double end, int num) {
-        if (num <= 0) return new double[0];
-        double[] out = new double[num];
+    private static float[] LinSpell(float start, float end, int num) {
+        if (num <= 0) return new float[0];
+        float[] out = new float[num];
         if (num == 1) {
             out[0] = start;
             return out;
         }
-        double step = (end - start) / (num - 1);
+        float step = (end - start) / (num - 1);
         for (int i = 0; i < num; i++) {
             out[i] = start + step * i;
         }
         return out;
     }
 
-    static double minArrival = -90.0;
-    static double maxArrival = -40.0;
+    static float minArrival = -90.0f;
+    static float maxArrival = -40.0f;
 
-    static double[] speeds = LinSpell(8.0, 12.0, 20);
-    static double[] launchAngles = LinSpell(40.0, 85.0, 45);
+    static float[] speeds = LinSpell(8.0f, 12.0f, 20);
+    static float[] launchAngles = LinSpell(40.0f, 85.0f, 45);
 
-    public static List<BallisticCalculatorResult> calculateForFuel(Translation3d shooterToTarget, double minArrival, double maxArrival, BallisticCalculatorMode mode) {
+    public static List<BallisticCalculatorResult> calculateForFuel(Translation3d shooterToTarget, float minArrival, float maxArrival, BallisticCalculatorMode mode) {
         return findConstrainedLaunch(
-                shooterToTarget.getX(),
-                shooterToTarget.getY(),
-                0.21,
-                0.11,
+                (float)shooterToTarget.getX(),
+                (float)shooterToTarget.getY(),
+                0.21f,
+                0.11f,
                 speeds,
                 launchAngles,
                 minArrival,
@@ -63,12 +63,12 @@ public class BallisticCalculator {
         );
     }
 
-    public static List<BallisticCalculatorResult> calculateForFuel(Translation3d shooterToTarget, double minArrival, double maxArrival) {
+    public static List<BallisticCalculatorResult> calculateForFuel(Translation3d shooterToTarget, float minArrival, float maxArrival) {
         return findConstrainedLaunch(
-                shooterToTarget.getX(),
-                shooterToTarget.getY(),
-                0.21,
-                0.11,
+                (float)shooterToTarget.getX(),
+                (float)shooterToTarget.getY(),
+                0.21f,
+                0.11f,
                 speeds,
                 launchAngles,
                 minArrival,
@@ -79,10 +79,10 @@ public class BallisticCalculator {
 
     public static List<BallisticCalculatorResult> calculateForFuel(Translation3d shooterToTarget) {
         return findConstrainedLaunch(
-                shooterToTarget.getX(),
-                shooterToTarget.getY(),
-                0.21,
-                0.11,
+                (float)shooterToTarget.getX(),
+                (float)shooterToTarget.getY(),
+                0.21f,
+                0.11f,
                 speeds,
                 launchAngles,
                 minArrival,
@@ -93,10 +93,10 @@ public class BallisticCalculator {
 
     public static List<BallisticCalculatorResult> calculateForFuel(Translation3d shooterToTarget, BallisticCalculatorMode mode) {
         return findConstrainedLaunch(
-                shooterToTarget.getX(),
-                shooterToTarget.getY(),
-                0.21,
-                0.11,
+                (float)shooterToTarget.getX(),
+                (float)shooterToTarget.getY(),
+                0.21f,
+                0.11f,
                 speeds,
                 launchAngles,
                 minArrival,
@@ -106,10 +106,10 @@ public class BallisticCalculator {
     }
 
     private static List<BallisticCalculatorResult> findConstrainedLaunch(
-            double targetX, double targetY,
-            double weightKg, double radiusM,
-            double[] vRange, double[] angleRange,
-            double a1deg, double a2deg,
+            float targetX, float targetY,
+            float weightKg, float radiusM,
+            float[] vRange, float[] angleRange,
+            float a1deg, float a2deg,
             BallisticCalculatorMode mode
     ) {
         if (weightKg <= 0 || radiusM <= 0) {
@@ -119,33 +119,33 @@ public class BallisticCalculator {
             throw new IllegalArgumentException("`v_range` and `angle_range` must be non-empty.");
         }
 
-        final double g = 9.81;
-        final double rho = 1.225;
-        final double Cd = 0.48;
-        final double area = Math.PI * radiusM * radiusM;
-        final double dragConst = 0.5 * rho * Cd * area;
-        final double eps = 1e-8;
-        final double errThresholdSq = 0.02 * 0.02;
+        final float g = 9.81f;
+        final float rho = 1.225f;
+        final float Cd = 0.48f;
+        final float area = (float)Math.PI * radiusM * radiusM;
+        final float dragConst = 0.5f * rho * Cd * area;
+        final float eps = 1e-8f;
+        final float errThresholdSq = 0.02f * 0.02f;
 
-        double aLow = Math.min(a1deg, a2deg);
-        double aHigh = Math.max(a1deg, a2deg);
+        float aLow = Math.min(a1deg, a2deg);
+        float aHigh = Math.max(a1deg, a2deg);
 
         final int nV = vRange.length;
         final int nA = angleRange.length;
         final int total = nV * nA;
 
         // precompute trig and flattened initial velocity components
-        double[] cosA = new double[nA];
-        double[] sinA = new double[nA];
+        float[] cosA = new float[nA];
+        float[] sinA = new float[nA];
         for (int i = 0; i < nA; i++) {
-            double rad = Math.toRadians(angleRange[i]);
-            cosA[i] = Math.cos(rad);
-            sinA[i] = Math.sin(rad);
+            float rad = (float)Math.toRadians(angleRange[i]);
+            cosA[i] = (float)Math.cos(rad);
+            sinA[i] = (float)Math.sin(rad);
         }
-        double[] vx0 = new double[total];
-        double[] vy0 = new double[total];
+        float[] vx0 = new float[total];
+        float[] vy0 = new float[total];
         for (int iv = 0; iv < nV; iv++) {
-            double v0 = vRange[iv];
+            float v0 = vRange[iv];
             int base = iv * nA;
             for (int ia = 0; ia < nA; ia++) {
                 vx0[base + ia] = v0 * cosA[ia];
@@ -157,49 +157,49 @@ public class BallisticCalculator {
         ConcurrentLinkedQueue<BallisticCalculatorResult> ballisticCalculatorResults = new ConcurrentLinkedQueue<>();
 
         // simulation parameters: coarse then fine
-        final double coarseDt = mode.getCoarseDt();
+        final float coarseDt = mode.getCoarseDt();
         final int coarseMaxSteps = 800; // coarse sweep
-        final double coarseThresholdSq = 0.06 * 0.06; // if coarse best below this, refine
-        final double fineDt = mode.getFineDt();
+        final float coarseThresholdSq = 0.06f * 0.06f; // if coarse best below this, refine
+        final float fineDt = mode.getFineDt();
         final int fineMaxSteps = 2000; // refinement window (starting from saved coarse state)
-        final double xMargin = 0.5;
-        final double yFloor = -1.0;
+        final float xMargin = 0.5f;
+        final float yFloor = -1.0f;
 
         IntStream.range(0, total).parallel().forEach(idx -> {
-            double initVx = vx0[idx];
-            double initVy = vy0[idx];
+            float initVx = vx0[idx];
+            float initVy = vy0[idx];
             int iv = idx / nA;
             int ia = idx % nA;
-            double v0 = vRange[iv];
-            double launchDeg = angleRange[ia];
+            float v0 = vRange[iv];
+            float launchDeg = angleRange[ia];
 
             // Quick bounding: if initial vx is tiny and targetX far, skip
             if (Math.abs(initVx) < eps && Math.abs(targetX) > 1.0) return;
 
             // ---- coarse pass ----
-            double vx = initVx;
-            double vy = initVy;
-            double x = 0.0;
-            double y = 0.0;
-            double bestErrSqCoarse = Double.POSITIVE_INFINITY;
-            double savedVx = vx, savedVy = vy, savedX = x, savedY = y;
+            float vx = initVx;
+            float vy = initVy;
+            float x = 0.0f;
+            float y = 0.0f;
+            float bestErrSqCoarse = Float.POSITIVE_INFINITY;
+            float savedVx = vx, savedVy = vy, savedX = x, savedY = y;
 
             for (int step = 0; step < coarseMaxSteps; step++) {
-                double v2 = vx * vx + vy * vy;
-                double v = Math.sqrt(v2);
-                double drag = dragConst * v2;
+                float v2 = vx * vx + vy * vy;
+                float v = (float)Math.sqrt(v2);
+                float drag = dragConst * v2;
 
-                double ax = (v > eps) ? -(drag * (vx / v)) / weightKg : 0.0;
-                double ay = (v > eps) ? -g - (drag * (vy / v)) / weightKg : -g;
+                float ax = (v > eps) ? -(drag * (vx / v)) / weightKg : 0.0f;
+                float ay = (v > eps) ? -g - (drag * (vy / v)) / weightKg : -g;
 
                 vx += ax * coarseDt;
                 vy += ay * coarseDt;
                 x += vx * coarseDt;
                 y += vy * coarseDt;
 
-                double dx = x - targetX;
-                double dy = y - targetY;
-                double errSq = dx * dx + dy * dy;
+                float dx = x - targetX;
+                float dy = y - targetY;
+                float errSq = dx * dx + dy * dy;
                 if (errSq < bestErrSqCoarse) {
                     bestErrSqCoarse = errSq;
                     savedVx = vx;
@@ -218,28 +218,28 @@ public class BallisticCalculator {
             vy = savedVy;
             x = savedX;
             y = savedY;
-            double bestErrSq = Double.POSITIVE_INFINITY;
-            double arrivalAngleDeg = Double.NaN;
+            float bestErrSq = Float.POSITIVE_INFINITY;
+            float arrivalAngleDeg = Float.NaN;
 
             for (int step = 0; step < fineMaxSteps; step++) {
-                double v2 = vx * vx + vy * vy;
-                double v = Math.sqrt(v2);
-                double drag = dragConst * v2;
+                float v2 = vx * vx + vy * vy;
+                float v = (float)Math.sqrt(v2);
+                float drag = dragConst * v2;
 
-                double ax = (v > eps) ? -(drag * (vx / v)) / weightKg : 0.0;
-                double ay = (v > eps) ? -g - (drag * (vy / v)) / weightKg : -g;
+                float ax = (v > eps) ? -(drag * (vx / v)) / weightKg : 0.0f;
+                float ay = (v > eps) ? -g - (drag * (vy / v)) / weightKg : -g;
 
                 vx += ax * fineDt;
                 vy += ay * fineDt;
                 x += vx * fineDt;
                 y += vy * fineDt;
 
-                double dx = x - targetX;
-                double dy = y - targetY;
-                double errSq = dx * dx + dy * dy;
+                float dx = x - targetX;
+                float dy = y - targetY;
+                float errSq = dx * dx + dy * dy;
                 if (errSq < bestErrSq) {
                     bestErrSq = errSq;
-                    arrivalAngleDeg = Math.toDegrees(Math.atan2(vy, vx));
+                    arrivalAngleDeg = (float)Math.toDegrees(Math.atan2(vy, vx));
                 }
 
                 if (y < yFloor || x > targetX + xMargin || (Math.abs(vx) < eps && Math.abs(vy) < eps)) {
@@ -248,8 +248,8 @@ public class BallisticCalculator {
             }
 
             if (!Double.isNaN(arrivalAngleDeg) && arrivalAngleDeg >= aLow && arrivalAngleDeg <= aHigh && bestErrSq < errThresholdSq) {
-                double roundedArrival = Math.round(arrivalAngleDeg * 100.0) / 100.0;
-                double bestErr = Math.sqrt(bestErrSq);
+                float roundedArrival = Math.round(arrivalAngleDeg * 100.0f) / 100.0f;
+                float bestErr = (float)Math.sqrt(bestErrSq);
                 ballisticCalculatorResults.add(new BallisticCalculatorResult(v0, launchDeg, roundedArrival, bestErr));
             }
         });
